refactor: View 폴더 생성
