+++ __tests__/BridgeSizeInputTest.js
@@ -1,4 +1,4 @@
-const InputView = require("../src/InputView");
+const InputView = require("../src/View/InputView");
 
 describe("다리 길이 입력값 유효성 검증 테스트", () => {
   it("다리 길이가 3~20 이외의 숫자로 입력될 경우 예외처리", () => {
