+++ __tests__/DirectionInputTest.js
@@ -1,4 +1,4 @@
-const InputView = require("../src/InputView");
+const InputView = require("../src/View/InputView");
 
 describe("다리 방향 유효성 검증 테스트", () => {
   it("U, D가 아닌 값이 입력될 경우 예외처리", () => {
