+++ src/GameController.js
@@ -1,8 +1,8 @@
 const { DIRECTION_KEY } = require("./constants/rule.js");
 
 const { Console } = require("@woowacourse/mission-utils");
-const InputView = require("./InputView.js");
-const OutputView = require("./OutputView.js");
+const InputView = require("./View/InputView.js");
+const OutputView = require("./View/OutputView.js");
 const BridgeGame = require("./BridgeGame.js");
 const Validation = require("./Validation.js");
