+++ src/View/OutputView.js
@@ -1,6 +1,6 @@
 const { Console } = require("@woowacourse/mission-utils");
-const { GAME_MSG } = require("./constants/message.js");
-const MapMaker = require("./MapMaker");
+const { GAME_MSG } = require("../constants/message.js");
+const MapMaker = require("../MapMaker");
 
 const OutputView = {
   printStart() {
